Implements subscription toggle and channel listing
Adds functionality to toggle user subscriptions to channels and retrieve lists of subscribers for a channel, as well as channels a user is subscribed to. Includes validation of channel and subscriber IDs, and handles cases where subscriptions are added or removed. Also fixes route assignments for subscriber and channel lists.